test_scripts/requirements.txt
google-genai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
openai>=1.40.0
gallery-dl>=1.31.0
requests>=2.31.0
yt-dlp>=2025.1.0
